# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register/initiate": {
      "post": {
        "operationId": "AuthController_registerInitiate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateRegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Verification OTP sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 201,
                    "message": "Registration initiated. Please check your email for the OTP verification code.",
                    "data": {
                      "tempUserId": "550e8400-e29b-41d4-a716-446655440000",
                      "email": "user@example.com"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "409": {
            "description": "Email address already in use"
          }
        },
        "summary": "Step 1: Initiate registration, generates a pending user and sends an email verification OTP",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/register/resend-otp": {
      "post": {
        "operationId": "AuthController_resendOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New verification OTP sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "New verification OTP sent successfully",
                    "data": null
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          }
        },
        "summary": "Step 1.5: Resend a fresh email verification OTP code",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/register/verify-otp": {
      "post": {
        "operationId": "AuthController_registerVerifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRegisterOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP successfully verified",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "OTP successfully verified",
                    "data": {
                      "verificationToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                      "expiresIn": 3600
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired OTP"
          }
        },
        "summary": "Step 2: Verify the registration OTP code",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/register/complete": {
      "post": {
        "operationId": "AuthController_registerComplete",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteRegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account password configured successfully. User must manually login.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Registration complete. You can now login.",
                    "data": {
                      "userId": "550e8400-e29b-41d4-a716-446655440000",
                      "email": "user@example.com"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Password confirmation mismatch"
          },
          "401": {
            "description": "Invalid or expired email verification code"
          }
        },
        "summary": "Step 3: Complete registration by setting the user password (manual login required after)",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Returns access and refresh tokens for normal users. If the password is system-generated and must be changed, no auth tokens are issued; the response includes passwordChangeToken and nextStep details for POST /auth/set-required-password.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login. Either tokens are returned, or passwordChangeToken is returned when requiresPasswordChange is true.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "example": {
                        "success": true,
                        "message": "Login successful",
                        "data": {
                          "tokens": {
                            "accessToken": "jwt-access-token",
                            "refreshToken": "jwt-refresh-token"
                          },
                          "user": {
                            "id": "user-id",
                            "email": "user@example.com",
                            "username": "user",
                            "mustChangePassword": false,
                            "roles": [
                              "User"
                            ],
                            "permissions": []
                          },
                          "requiresPasswordChange": false,
                          "message": "Login successful"
                        }
                      }
                    },
                    {
                      "example": {
                        "success": true,
                        "message": "Login successful",
                        "data": {
                          "success": false,
                          "requiresPasswordChange": true,
                          "passwordChangeToken": "short-lived-token",
                          "nextStep": {
                            "action": "SET_PASSWORD",
                            "method": "POST",
                            "endpoint": "/auth/set-required-password"
                          },
                          "message": "Temporary password accepted. Set a new password, then log in again with the new password."
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or account inactive"
          }
        },
        "summary": "Log in with username or email and password",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/set-required-password": {
      "post": {
        "description": "Consumes the short-lived passwordChangeToken returned by login when requiresPasswordChange is true. This endpoint sets the real password, clears mustChangePassword, revokes existing sessions/tokens, and does not issue auth tokens. The user must log in again with the new password.",
        "operationId": "AuthController_setRequiredPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRequiredPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password set successfully. User must login again.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Password set successfully. Please login again.",
                    "data": {
                      "success": true,
                      "message": "Password set successfully. Please log in again with your new password."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Password confirmation mismatch or validation failed"
          },
          "401": {
            "description": "Invalid or expired password change token"
          }
        },
        "summary": "Set a new password after temporary-password login, then require normal login",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/switch-tenant": {
      "post": {
        "operationId": "AuthController_switchTenant",
        "parameters": [],
        "responses": {
          "200": {
            "description": "New token pair generated for the requested tenant",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Workspace switched successfully",
                    "data": {
                      "accessToken": "jwt-access-token",
                      "refreshToken": "jwt-refresh-token"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized or not a member of the tenant"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Switch the active tenant context for this session",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New token pair generated",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Tokens successfully refreshed",
                    "data": {
                      "accessToken": "jwt-access-token",
                      "refreshToken": "jwt-refresh-token"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid/expired token or reuse security violation"
          }
        },
        "summary": "Rotate JWT access and refresh tokens",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Session revoked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 204,
                    "message": "Successfully logged out",
                    "data": null
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Log out from the current active session",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification OTP generated",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Password recovery request initiated",
                    "data": {
                      "verificationId": "550e8400-e29b-41d4-a716-446655440000",
                      "email": "user@example.com"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Step 1: Request password recovery using username or email address",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/reset-password/verify": {
      "post": {
        "operationId": "AuthController_resetPasswordVerify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Password recovery OTP verified successfully",
                    "data": {
                      "resetToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                      "expiresIn": 3600
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired OTP"
          }
        },
        "summary": "Step 2: Verify the password recovery OTP code",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "SaaS Multi-Tenant API",
    "description": "Auto-generated OpenAPI specification for the SaaS Demo application.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token to authenticate",
        "in": "header"
      }
    },
    "schemas": {
      "InitiateRegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "johndoe@example.com",
            "description": "Valid email address of the user"
          },
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "First name of the user"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "Last name of the user"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResendOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "johndoe@example.com",
            "description": "Valid email address associated with the pending registration"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyRegisterOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "johndoe@example.com",
            "description": "Email address of the user initiating registration"
          },
          "otp": {
            "type": "string",
            "example": "123456",
            "description": "6-digit OTP code sent to the email"
          }
        },
        "required": [
          "email",
          "otp"
        ]
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "example": "johndoe",
            "description": "Username or email address of the user"
          },
          "password": {
            "type": "string",
            "example": "SuperSecurePassword123!",
            "description": "User password"
          }
        },
        "required": [
          "identifier",
          "password"
        ]
      },
...
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "Valid refresh token"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
    }
  }
}